requirement.txt
numpy==1.32.5
scikit-learn==1.6.1
pandas==2.2.3
tensorflow==2.10.0
matplotlib==3.10.0
seaborn
flask==3.1.0
pillow==23.2.1
